requirements.txt
django>=4.2.3
pytest-django>=4.5.2
djangorestframework>=3.14.0
python-dotenv>=1.0.0
pika>=1.3.2
celery>=5.3.1
feedparser>=6.0.10
httpx>=0.24.1
beautifulsoup4>=4.12.2-py3-none-any.whl
lxml>=4.9.3
